+++ src/components/manage/Blocks/Hero/Data.jsx
@@ -3,10 +3,11 @@ import { useIntl } from 'react-intl';
 import schemaHero from './schema.js';
 import { BlockDataForm } from '@plone/volto/components';
 
-const HeroImageLeftBlockData = (props) => {
+const HeroBlockData = (props) => {
   const { block, data, onChangeBlock } = props;
   const intl = useIntl();
   const schema = schemaHero({ ...props, intl });
+
   return (
     <BlockDataForm
       schema={schema}
@@ -24,4 +25,4 @@ const HeroImageLeftBlockData = (props) => {
   );
 };
 
-export default HeroImageLeftBlockData;
+export default HeroBlockData;