+++ src/components/manage/Blocks/Hero/Edit.jsx
@@ -1,0 +1,215 @@
+/**
+ * Edit Hero block.
+ */
+
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import { compose } from 'redux';
+import { readAsDataURL } from 'promise-file-reader';
+import { isEqual } from 'lodash';
+import { defineMessages, injectIntl } from 'react-intl';
+import cx from 'classnames';
+
+import { getBaseUrl } from '@plone/volto/helpers';
+import { createContent } from '@plone/volto/actions';
+import { SidebarPortal, Hero } from '@plone/volto/components';
+import { withBlockExtensions } from '@plone/volto/helpers';
+
+import Data from './Data';
+
+const messages = defineMessages({
+  title: {
+    id: 'Title',
+    defaultMessage: 'Title',
+  },
+  description: {
+    id: 'Description',
+    defaultMessage: 'Description',
+  },
+  placeholder: {
+    id: 'Upload a new image',
+    defaultMessage: 'Upload a new image',
+  },
+  image: {
+    id: 'Image',
+    defaultMessage: 'Image',
+  },
+  browse: {
+    id: 'Browse',
+    defaultMessage: 'Browse',
+  },
+  uploading: {
+    id: 'Uploading image',
+    defaultMessage: 'Uploading image',
+  },
+});
+
+/**
+ * Edit image block class.
+ * @class Edit
+ * @extends Component
+ */
+class EditComponent extends Component {
+  /**
+   * Property types.
+   * @property {Object} propTypes Property types.
+   * @static
+   */
+  static propTypes = {
+    selected: PropTypes.bool.isRequired,
+    block: PropTypes.string.isRequired,
+    index: PropTypes.number.isRequired,
+    data: PropTypes.objectOf(PropTypes.any).isRequired,
+    content: PropTypes.objectOf(PropTypes.any),
+    request: PropTypes.shape({
+      loading: PropTypes.bool,
+      loaded: PropTypes.bool,
+    }).isRequired,
+    pathname: PropTypes.string.isRequired,
+    onChangeBlock: PropTypes.func.isRequired,
+    onSelectBlock: PropTypes.func.isRequired,
+    onDeleteBlock: PropTypes.func.isRequired,
+    onFocusPreviousBlock: PropTypes.func.isRequired,
+    onFocusNextBlock: PropTypes.func.isRequired,
+    handleKeyDown: PropTypes.func.isRequired,
+    createContent: PropTypes.func.isRequired,
+    editable: PropTypes.bool,
+  };
+
+  /**
+   * Default properties
+   * @property {Object} defaultProps Default properties.
+   * @static
+   */
+  static defaultProps = {
+    editable: true,
+  };
+
+  /**
+   * Constructor
+   * @method constructor
+   * @param {Object} props Component properties
+   * @constructs WysiwygEditor
+   */
+  constructor(props) {
+    super(props);
+
+    this.onUploadImage = this.onUploadImage.bind(this);
+    this.state = {
+      uploading: false,
+    };
+  }
+
+  /**
+   * Component will receive props
+   * @method componentWillReceiveProps
+   * @param {Object} nextProps Next properties
+   * @returns {undefined}
+   */
+  UNSAFE_componentWillReceiveProps(nextProps) {
+    if (
+      this.props.request.loading &&
+      nextProps.request.loaded &&
+      this.state.uploading
+    ) {
+      this.setState({
+        uploading: false,
+      });
+      this.props.onChangeBlock(this.props.block, {
+        ...this.props.data,
+        url: nextProps.content['@id'],
+      });
+    }
+  }
+
+  /**
+   * @param {*} nextProps
+   * @param {*} nextState
+   * @returns {boolean}
+   * @memberof Edit
+   */
+  shouldComponentUpdate(nextProps) {
+    return this.props.selected || !isEqual(this.props.data, nextProps.data);
+  }
+
+  /**
+   * Upload image handler
+   * @method onUploadImage
+   * @returns {undefined}
+   */
+  onUploadImage({ target }) {
+    const file = target.files[0];
+    this.setState({
+      uploading: true,
+    });
+    readAsDataURL(file).then((data) => {
+      const fields = data.match(/^data:(.*);(.*),(.*)$/);
+      this.props.createContent(
+        getBaseUrl(this.props.pathname),
+        {
+          '@type': 'Image',
+          image: {
+            data: fields[3],
+            encoding: fields[2],
+            'content-type': fields[1],
+            filename: file.name,
+          },
+        },
+        this.props.block,
+      );
+    });
+  }
+
+  /**
+   * Render method.
+   * @method render
+   * @returns {string} Markup for the component.
+   */
+  render() {
+    if (__SERVER__) {
+      return <div />;
+    }
+    const placeholder =
+      this.props.data.placeholder ||
+      this.props.intl.formatMessage(messages.placeholder);
+
+    return (
+      <div
+        className={cx(
+          'block hero align',
+          {
+            selected: this.props.selected,
+          },
+          {
+            center: !Boolean(this.props.data.align),
+          },
+          this.props.data.align,
+        )}
+      >
+        <Hero
+          {...this.props}
+          uploading={this.state.uploading}
+          placeholder={placeholder}
+          onUploadImage={this.onUploadImage}
+          isEditMode
+        />
+        <SidebarPortal selected={this.props.selected}>
+          <Data {...this.props} />
+        </SidebarPortal>
+      </div>
+    );
+  }
+}
+
+export default compose(
+  injectIntl,
+  withBlockExtensions,
+  connect(
+    (state, ownProps) => ({
+      request: state.content.subrequests[ownProps.block] || {},
+      content: state.content.subrequests[ownProps.block]?.data,
+    }),
+    { createContent },
+  ),
+)(EditComponent);