+++ src/components/manage/Blocks/Hero/components/index.js
@@ -1,0 +1,3 @@
+export HeroImage from './HeroImage';
+export HeroText from './HeroText';
+export HeroToolbar from './HeroToolbar';